Finalise component organisation
