+++ src/components/Screens/NewStageScreen/Wizard.js
