+++ src/components/Screens/NewStageScreen/index.js
