+++ src/styles/_components.scss
@@ -25,6 +25,7 @@
 @import './components/list';
 @import './components/new-stage';
 @import './components/network';
+@import './components/new-stage-screen';
 @import './components/options';
 @import './components/overview';
 @import './components/panel';
@@ -38,7 +39,6 @@
 @import './components/simple-list';
 @import './components/stage-editor';
 @import './components/stage-heading';
-@import './components/stage-type-selector';
 @import './components/start';
 @import './components/thumbnail';
 @import './components/timeline';
