+++ src/styles/components/_new-stage-screen.scss
