--- src/styles/components/_new-stage.scss
@@ -1,5 +1,0 @@
-@import './new-stage/new-stage';
-@import './new-stage/navigation';
-@import './new-stage/interface-category';
-@import './new-stage/interface';
-
