--- src/styles/components/new-stage/_interface-category.scss
@@ -1,21 +1,0 @@
-.new-stage-interface-category {
-  border-left-color: transparent;
-  border-left-width: unit(1);
-  border-left-style: solid;
-  padding: unit(4) 0;
-
-  &__title {
-    margin-top: 0;
-    margin-bottom: unit(4);
-  }
-
-  &__interfaces {
-    grid-template-columns: 1fr 1fr 1fr;
-    grid-gap: unit(4);
-    display: grid;
-  }
-
-  &__interface {
-    display: inline-flex;
-  }
-}
