--- src/styles/components/new-stage/_interface.scss
@@ -1,54 +1,0 @@
-.new-stage-interface {
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-
-  &__title {
-    margin: 0;
-  }
-
-  &__preview {
-    margin-top: 1rem;
-    width: 100%;
-    min-width: unit(30);
-    border-width: .4rem;
-    border-style: solid;
-    border-color: var(--background-dark);
-    background-color: var(--background-dark);
-    transition-property: border-color;
-    transition-duration: var(--animation-duration-standard);
-    transition-timing-function: var(--animation-easing);
-    border-radius: .75rem;
-  }
-
-  &__tags {
-    margin-top: 1rem;
-  }
-
-  &__tag {
-    background: transparent;
-    color: var(--text-light);
-    display: inline-block;
-    font-size: .8rem;
-    padding: 0 .5rem;
-    height: 1.3rem;
-    line-height: 1.3rem;
-    border-radius: 1.3rem;
-
-    @each $category in $new-stage-categories {
-      &--#{$category} {
-        background: var(--color-#{$category});
-      }
-    }
-  }
-
-  &:hover {
-    .new-stage-interface {
-      &__preview {
-        border-color: var(--architect-active);
-      }
-    }
-  }
-}
