--- src/styles/components/new-stage/_navigation.scss
@@ -1,27 +1,0 @@
-.new-stage-navigation {
-  background: var(--color-slate-blue--dark);
-  display: flex;
-  color: var(--text-light);
-  padding: 0 2rem;
-  border-radius: var(--border-radius) var(--border-radius) 0 0;
-
-  &__category {
-    padding: 0 1rem;
-    height: 5rem;
-    line-height: 5rem;
-    background-color: transparent;
-    cursor: pointer;
-    transition-property: background-color;
-    transition-duration: var(--animation-duration-standard);
-    transition-timing-function: var(--animation-easing);
-
-    @each $category in $new-stage-categories {
-      &--#{$category} {
-        &:hover {
-          background-color: var(--color-#{$category});
-          color: var(--text-light);
-        }
-      }
-    }
-  }
-}
