--- src/styles/components/new-stage/_new-stage.scss
@@ -1,32 +1,0 @@
-$new-stage-categories: [name-generators name-interpreters sociograms utilities];
-
-.new-stage {
-  --color-name-generators: var(--ord-color-seq-2);
-  --color-sociograms: var(--ord-color-seq-3);
-  --color-name-interpreters: var(--ord-color-seq-4);
-  --color-utilities: var(--ord-color-seq-5);
-
-  width: 100%;
-  padding: 0 0 4rem;
-
-  &__navigation {
-    position: sticky;
-    width: 100%;
-    top: 0;
-    left: 0;
-  }
-
-  &__category {
-    padding: unit(2) unit(4) ;
-    background: var(--architect-panel-grey);
-
-    &:last-child {
-      border-radius: 0 0 var(--border-radius) var(--border-radius);
-    }
-  }
-
-  &__options {
-    max-width: 80rem;
-    margin: 0 auto;
-  }
-}
